chore: #91 move txn setters to separate file
