+++ src/ledger/txn/mod.rs
@@ -1,4 +1,5 @@
 pub mod getters;
+pub mod setters;
 pub mod utils;
 // imports
 use chrono::prelude::*;
@@ -158,47 +159,4 @@ impl Txn {
 
         txn
     }
-
-    /////////////////////////////////////////////////////////////////////
-    ////////////////////////////// SETTERS //////////////////////////////
-
-    /// ## Get identifier (hash) for txn and set on txn object and store the output on the Txn object
-    ///
-    /// Returns id
-    pub fn set_id(&mut self) -> TxnId {
-        let id = self.calc_id();
-        self.id = Some(id);
-
-        id
-    }
-    /// ## Set the signature for the transaction.
-    pub fn set_signature(&mut self, signature: TxnSignature) {
-        self.signature = Some(signature);
-    }
-
-    /// ## Add the signature to the transaction body in place.
-    ///
-    /// 1) Sign the transaction
-    /// 2) Add signature to transaction body
-    /// 3) Return signature
-    pub fn sign(&mut self, wallet: &Wallet) -> TxnSignature {
-        let sig = self.calc_signature(wallet);
-        self.set_signature(sig.to_owned());
-
-        sig
-    }
-
-    ////////////////////////////// SETTERS //////////////////////////////
-    /////////////////////////////////////////////////////////////////////
-
-    /////////////////////////////////////////////////////////////////////
-    /////////////////////////////// UTILS ///////////////////////////////
-
-    /////////////////////////////// UTILS ///////////////////////////////
-    /////////////////////////////////////////////////////////////////////
-
-    /////////////////////////////////////////////////////////////////////
-    ///////////////////////////// VALIDATION ////////////////////////////
-    ///////////////////////////// VALIDATION ////////////////////////////
-    /////////////////////////////////////////////////////////////////////
 }
