+++ src/ledger/txn/setters.rs
@@ -1,0 +1,17 @@
+use super::{Txn, TxnId};
+use crate::utils::signature::TxnSignature;
+
+impl Txn {
+    /// ## Get identifier (hash) for txn and set on txn object and store the output on the Txn object
+    /// Returns id.
+    pub fn set_id(&mut self) -> TxnId {
+        let id = self.calc_id();
+        self.id = Some(id);
+
+        id
+    }
+    /// ## Set the signature for the transaction.
+    pub fn set_signature(&mut self, signature: TxnSignature) {
+        self.signature = Some(signature);
+    }
+}
